WIP: Upgrade gwtjsonrpc to 1.12
This version of gwtjsonrpc includes a fix for encoding/decoding tokens
with the '+' character.

Add a test to confirm that it behaves as expected. The test is based on
the one added by Teng Long in change Ib81f1440b.

TODO: The WORKSPACE needs to be updated to use the released gwtjsonrpc
artifact. For now it is necessary to build/install it locally, which will
most likely require to manually change the sha1.

Bug: Issue 12424
Change-Id: Iad98efe04ca43d8e2d8186082e2718bf1dfd0e79