Update patch set 7
Patch Set 7: Code-Review+2

Whoa, awesome! Thanks Dmitrii!

I did not comment on language and on nits, because I think in this case it is not worth to spend time on bikeshedding too much. Overall I could understand everything just fine. And my understanding is that we will be replacing WCT later this year anyway, right? That means most of the code in this change in temporary and does not need to be perfect. :-)

Patch-set: 7
Reviewer: Gerrit User 1013302 <1013302@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=+2