Make cache disk stat metric computation optional
Computing the disc stat metrics for persistent caches is very expensive
and can block a thread for several minutes on larger installations.

This change adds a configuration option to enable the computation of
the disk stats. This option is set to false by default. This makes the
change a breaking change, which will affect users that use the
caches_disk_.* metrics. However, since computing this metric can block
multiple threads depending on metric scrape rate and thus even cause
outages, it makes sense to not enable this by default.

Change-Id: I41ee2d9a368c312b7b2729d17d6c19bee0d90922