Fix account-link breaking onto a second line
The issue is exposed by the recent change 257912, which allows reviewers
being shown next to each other, which saves vertical space, especially
when first names are used as a display name. For this to the
<gr-reviewer-list> element needs to set a max-width, and if the account
element does not fit into this max-width, then we are seeing ugly line
breaks introduced.

To fix this we want to apply a set of css rules that used to be set in
gr-change-metadata_html.js. And these rules were always meant to apply
also to reviewers, but my guess is that this broke with the
introduction of Shadow DOM, but nobody notcied that.

We do want to keep the ability to not clip account elements (for example
in the dashboard), but also don't want to repeat the set of css rules
everywhere for nice clipping. So we are moving the rules into
<gr-account-link>, but are setting a very high default for max-width by
default. For limiting the length we introduce the `--account-max-length`
css variable and set that to 20ch for all account widgets that are
children of <gr-change-metadata>.

Change-Id: I036701d46a45131189fc00f1f189c1314d586a87