Fix 'visibleto' predicate for group
Issue:
When one uses a valid group name in 'visibleto' predicate no change
gets returned even if it should be from the access perspective.

Solution:
It boils down to the problem in ChangeIsVisibleToPredicate for
SingleGroupUser user - it is not an IdentifiedUser hence it gets
swapped for AnonymousUser and visibility check fails unless changes
in question are granted to Anonymous Users group. Fix it to let it
use the user it has in case it is SingleGroupUser or follow existing
user resolution (with fallback to anonymous) as it was before.

Bug: Issue 12606
Change-Id: Ieea8abf4b52d528d4cb2503270c7eff68476fc6c