Update patch set 4
Patch Set 4: Code-Review-1

The fix looks reasonable, but one of the Elasticsearch tests seems to be failing.  I'm not sure if that's related or it's just a flaky test.  Regardless, it would be good to also add an additional test that would expose the issue and prove that it's fixed.

Patch-set: 4
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1