Update patch set 10
Patch Set 10:

> Patch Set 8:
> 
> Feel free to ignore, but if you're accessing the singletons from a static appContext, what's the benefit that you're trying to get from even assigning them to fields at all? Is it just that `this.flags.foo` is shorter than `appContext.flags.foo`?
> 
> (Also wow, first time using Gerrit in recent times, and it's basically meme/5336480218873856)

Thanks for looking on our implementation Maksym. We follow Angular and other framework way of setting instance of service in constructor and then accessing it from variable. It's probably mostly based of how we used to use services. We will allow both ways (especially on static methods that needs service - we use just appContext.flags.foo), but prefer this.flags.

Patch-set: 10