Allow to configure app max size on large screens
One of the pain point of migrating from GWT to Polygerrit is the usage
of space: the way Polygerrit does so can be quite dispersive and loose
when compared to the compactness of the old GWT UI.

This is especially evident on big screens, where, due to its reactive
nature, Polygerrit spreads taking over all available space, thus leaving
lots of unused space in-between the content of a change, which makes it
more difficult to do code-review effectively.

This change allows to push unused space to the margins, so that the
content of any page is more compact.

By default the app is not constrained in size, however users can
override the CSS by tuning the relevant CSS variables. For example:

  --app-max-width: 1600px;
  --app-max-width-box-shadow: 0px 4px 8px 12px rgba(199,199,199,1);
  --app-max-width-border: solid 1px black;
  --app-max-width-background-color: var(--view-background-color);

Which will apply on screens larger than 100em (1600px)

Change-Id: Icff0c44487a48c74e00e58f7cdba37fe44b29b62