Allow to configure margins on large screens
One of the pain point of migrating from GWT to Polygerrit is the usage
of space: the way Polygerrit does so can be quite dispersive and loose
when compared to the compactness of the old GWT UI.

This is especially evident on big screens, where, due to its reactive
nature, Polygerrit spreads taking over all available space, thus leaving
lots of unused space in-between the content of a change, which makes it
more difficult to do code-review effectively.

This change allows to push unused space to the margins, so that the
content of any page is more compact.

By default margins are set to 0em, however users can override the CSS
by setting:

  --large-screen-left-margin: <some_value>
  --large-screen-right-margin <some_value>

Which will apply on screens larger than 80em (1280px)

Change-Id: Icff0c44487a48c74e00e58f7cdba37fe44b29b62