Allow resetting scroll without grabbing focus
It is sometimes necessary to resetting scroll behavior when a new diff
is loaded after the window scroll disabled scrolling, and one does not
necessarily want to set the cursor which also grabs focus.
This change provides a new method to do so.
It does not change anything about the behavior of Gerrit for now.

Change-Id: I64875ca271967baf970288aaebd19b96ddbb5542