Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > I have not tested it, but I could test it as well.
> > 
> > Would appreciate it.  Thanks.
> 
> Sorry for the delay, I am busy with other tasks. What exact the series of changes that should be built, published and verified and on what branches?

gwtjsonrpc 2.12-SNAPSHOT should be built from [1] and installed to the local maven repo, then this change should be modified locally to fix the sha1 to match the version of gwtjsonrpc that was installed.

once you've reviewed/tested it and confirmed it's OK, I'll go ahead and make the official release of gwtjsonrpc 2.12 and then update this change to use it.

for now this is only on stable-2.16. the gwtjsonrpc library is not used from stable-3.0 onwards, so we'll need to move Teng's other change from master to stable-3.0 and adapt it to match the changes I've done here.

[1] https://gerrit-review.googlesource.com/c/gwtjsonrpc/+/263415/2

Patch-set: 1