Always verify PolyGerrit if bazel related files are changed
Polygerrit was not verified, if Bazel related files were changed
outside of the polygerrit-ui directory except for the WORKSPACE
file. However, there are Bazel related files that are influencing
the build and tests of Polygerrit.

Thus, now if there was a change to a Bazel related file, the
Polygerrit UI will also be verified.

Change-Id: Ic217a7abcc384f170a9e6afb713b2b9c484d10aa