Fix appContext for shared gr-diff component
appContext contains services and different files can import appContext. Right
now it is hard to find all references to appContext from gr-diff.
Because we have only one service, we can mock it with empty service. In this
case any unexpected usages of flags service from gr-diff are not break gr-diff.

Change-Id: Ic0fc226d92502ac9a2e2370406847bd928a5f41f