Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Hey @Ben,
> > > thanks for your initial feedback on this 😊
> > > Any other comment anyone?
> > 
> > Hey Antonio, Delphine will get back to you tomorrow. I think she has some interesting experiences to share from trying something similar in another app.
> > 
> > Reading the commit message once more, I think it would be extremely helpful, if you could provide an example of a GWT UI and a Polygerrit UI page that expose the regression that you would like to fix. Otherwise I find it hard to grasp what the issue is exactly with being loose vs compact. It would also provide a better opportunity for us to suggest alternative ways to address the issue.
> 
> Hey Ben, yep absolutely, looking forward from Delphine's input on this 😊
> 
> Regarding the example, you're right, I had already commented on this change with two links on:
> * what I think is the problem in Polygerrit: (https://ibb.co/hKsNMcj)
> * the proposed solution (https://ibb.co/1stpWNn - https://www.youtube.com/watch?v=_h72ah6yzis)
> 
> Here is the missing GWT example: (https://ibb.co/gdt18QQ)
> 
> Please let me know if that's all, happy to add more info if needed.

Hi Antonio, first thank you for proposing to improve the UI and sharing screenshots and video. 

We agree that there is a lot of wasted whitespace in Gerrit. From my experience, we have tried a similar approach to set a max-width in our internal tools and got a lot of user pushbacks:

- Users were annoyed that so much space was wasted on the side on big screen sizes, while this space could have been used to show more information at a glance. Users complained about the bad layout and reducing screen-width did not improve the layout according to them.
- Users thought the UI was broken and did not appreciate control was removed from them to set the UI to a width they choose (resizing manually a browser window for example, fitting the full size of the monitor)
- Setting max-width also made the UI look broken because diff with more than 100char were “bleeding” in the empty space on the side

We ended up removing max-width and to use the entire width of the screen (as Gerrit does at the moment).

As a conclusion, the tools had more of a layout issue (general layout and diff display depending on diff width preferences and various screen widths) and there are similar issues in Gerrit. So that is why I don’t want to push this change before assessing how we could improve the layout in Gerrit and how we could improve displaying diffs, this will be an effort that we will take step by step as we work on the UX and UI of Gerrit in the following quarters.

Patch-set: 4
Reviewer: Gerrit User 1079742 <1079742@173816e5-2b9a-37c3-8a2e-48639d4f1153>