Update patch set 5
Patch Set 5:

> A quick trial cherry-pick revealed that it would be quite complex.

What stable branch have you tried? I would also say, that this is a non introusive change. This change depends only on some upgrades of third party dependencies:

* SSHD to 2.4
* JGit to most recent master

and is adding two new dependencies:

* JGit-ssh-apache from JGit submodule (that is already cloned anyway, but not used prior to this change)
* sshd-sftp that is transitive dependency of the above

Moreover, Francois Ferrand has already uploaded a similar SSHD upgrade to 2.16 stable branch: [1]. Pobably, upgrade JGit to most recent master on stable branches would be more involved?

Anyway, it would be already great, if we could try to submit this and provide that option in upcoming 3.2 release. It must be enabled in configuration an we could even declare it as experimental.

@DavidP, @Matthias what about upgrading the JGit to most recent master? Should we extract that part from this change and do it in own change (and rebase that change on top of it)?

As stated in the commit message that upgrade is needed, to align SSHD version used in Gerrit and JGit to 2.4.

* [1] https://gerrit-review.googlesource.com/c/gerrit/+/254658

Patch-set: 5
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>