Update patch set 4
Patch Set 4:

> Hi Antonio, first thank you for proposing to improve the UI and sharing screenshots and video. 
> 
> We agree that there is a lot of wasted whitespace in Gerrit. From my experience, we have tried a similar approach to set a max-width in our internal tools and got a lot of user pushbacks:

I believe Tony's proposal isn't about limiting the max-width, but rather adding a configurable left/right space based on the max screen size. By default it would be a noop for Gerrit and for your users.

Do you see a problem with your user-base? Would they notice any difference?

> - Users were annoyed that so much space was wasted on the side on big screen sizes, while this space could have been used to show more information at a glance. Users complained about the bad layout and reducing screen-width did not improve the layout according to them.

I believe in the example that Tony has shown there is a lot of white space wasted in the current UI. The information is spread across the wide screen (left and right), forcing the user to turn the eyes from left to right all the times. Users have raised concerns of tiredness of using it on a daily basis.

Do you have users with large panoramic screens? (the curved ones)
Do they have issues in having the file names on the left of the curve and their status on the right of the curve?

> - Users thought the UI was broken and did not appreciate control was removed from them to set the UI to a width they choose (resizing manually a browser window for example, fitting the full size of the monitor)
> - Setting max-width also made the UI look broken because diff with more than 100char were “bleeding” in the empty space on the side

Do you believe Tony's change would force a max-width on Gerrit UX?
Can you propose a way to improve Tony's proposal?

The health&safety issue raised on the user turning their eyes constantly left and right could be a showstopper. What do you think?

Patch-set: 4