Update patch set 5
Patch Set 5:

> > > As stated in the commit message that upgrade is needed, to align SSHD version used in Gerrit and JGit to 2.4.
> > 
> > That's actually an important point. As soon as Gerrit starts using org.eclipse.jgit.ssh.apache, we will have to coordinate updating the JGit org.apache.sshd version and the Gerrit org.apache.sshd version. Or at least the Gerrit people have to be aware that whenever they upgrade JGit, it might cause an upgrade of sshd, too. (Unfortunately the org.apache.sshd interfaces change in unpredictable ways between releases, so JGit has to requires a narrow version range - a particular implementation of org.eclipse.jgit.ssh.apache will in general only work with a specific org.apache.sshd version.)
> 
> There are existing libs that have hard dependencies on other libs.  We manage this by adding comments in the WORKSPACE file like "When X is upgraded, Y must also be upgraded"

Sure the versions can be managed. But IIRC some sshd updates in the past in Gerrit sometimes were not without problems? Gerrit so far uses sshd for the server side (ssh daemon); JGit uses it for the client side, and it tests basically only the client side. (Ok, there's a little test server, too, but that hardly can qualify as testing the sshd server-side.) Looks to me that we now would have to first make sure that an upgrade of sshd doesn't break a real server-side before we could safely upgrade the sshd dependency in org.eclipse.jgit.sshd.apache.

Or am I seeing problems where there are none? What do we do if I upgrade sshd in JGit to 2.6.0, everything looks fine, and then Gerrit upgrades to that JGit version and discovers that its sshd-daemon has a problem with 2.6.0?

Patch-set: 5