Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > > > As stated in the commit message that upgrade is needed, to align SSHD version used in Gerrit and JGit to 2.4.
> > > 
> > > That's actually an important point. As soon as Gerrit starts using org.eclipse.jgit.ssh.apache, we will have to coordinate updating the JGit org.apache.sshd version and the Gerrit org.apache.sshd version. Or at least the Gerrit people have to be aware that whenever they upgrade JGit, it might cause an upgrade of sshd, too. (Unfortunately the org.apache.sshd interfaces change in unpredictable ways between releases, so JGit has to requires a narrow version range - a particular implementation of org.eclipse.jgit.ssh.apache will in general only work with a specific org.apache.sshd version.)
> > 
> > There are existing libs that have hard dependencies on other libs.  We manage this by adding comments in the WORKSPACE file like "When X is upgraded, Y must also be upgraded"

But the JGit is consumed as git submodule, so that we should another way to document "This to upgrade SSH version if upgrading JGit".
 
> Sure the versions can be managed. But IIRC some sshd updates in the past in Gerrit sometimes were not without problems? Gerrit so far uses sshd for the server side (ssh daemon); JGit uses it for the client side, and it tests basically only the client side. (Ok, there's a little test server, too, but that hardly can qualify as testing the sshd server-side.) Looks to me that we now would have to first make sure that an upgrade of sshd doesn't break a real server-side before we could safely upgrade the sshd dependency in org.eclipse.jgit.sshd.apache.

Good point. Gerrit project has ESC and CM teams now, and we always wanted to intensify the cooperation between different projects in Gerrit community, so may be this is a perfect opportunity to start such cooperation?

> Or am I seeing problems where there are none? What do we do if I upgrade sshd in JGit to 2.6.0, everything looks fine, and then Gerrit upgrades to that JGit version and discovers that its sshd-daemon has a problem with 2.6.0?

Yeah, unfortunately.

What we are doing here, in this change, is actually introducing dependency hell/tight coupling between client and server parts of Gerrit product.

We had very similar situation with ElasticSearch integration in Gerrit core, that we escaped recently.

See my talk that I gave during Gerrit User Conference 2018, in Palo Alto, CA: [1].

The relevant slides for this discussion are: 14 - 19.

But the situaion with ES integration in Gerrit was less problematic.

Why? Because pulling ES backend als dependency in Gerrit core was a misconeption, wrong and not needed from the day one. And was only needed because the ES client implementation "Jest" depends on ES backend, that depends on Lucene. So to say ES server was pulled in release.war of gerrit, but the ES server part in release.war was dead (unused).

While that simplified ES integration tests, as ES node could be started directly from gerrit, it introduced tight coupling between ES and Lucene versions (gerrit uses Lucene as its own secondary index backend).

The solution was trivial, switch to using a different ES client library without dependency on ES backend (and Lucene) version. And entirely remove the ES backend from Gerrit core. That way Gerrit currently supports different ES versions: 5 and 6 (that depends on different version of Lucene).

However, the situation with SSH client/SSH daemon is much worse: we cannot externalize/remove SSH daemon part from the Gerrit server: it must bind the whole Guice universe of gerrit backend code. Moreover, we cannot extract the client because it is part of JGit that is the major dependeny of gerrit itself. And given that sshd-core and sshd-client artifacts are the same library that shares many parts in common, it cannot be de-coupled from each other, so that sshd-client 2.4 wouldn't probably work with sshd-server 3.0. (Even if they would work together we cannot put them on the same classpath, they have to be in release.war).

* [1] https://ostrovsky.org/gerrit/bazel-new-and-noteworthy/bazel-new-and-noteworthy.html

Patch-set: 5
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>