Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> Why do we need this major upgrade in a stable branch?
> Do we have any blocking issue? What is the issues that we are resolving?

I'm working on the plausibility of deploying the jgit-mina fix on stable-2.16.
Part of that is to upgrade Mina to 2.4 and to make this more clear and in an effort to minimize merge conflicts when/if this is submitted and merge into 3.0, 3.1, master I cherry-picked the 2.3 upgrade as an intermediary step towards the 2.4 upgrade (also a cherry-pick).
See comments at https://gerrit-review.googlesource.com/c/gerrit/+/263692

Patch-set: 1