Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > Why do we need this major upgrade in a stable branch?
> > Do we have any blocking issue? What is the issues that we are resolving?
> 
> I'm working on the plausibility of deploying the jgit-mina fix on stable-2.16.

Can you flag the changes as WIP or [DO NOT MERGE]? Otherwise they'll risk to get merged as I release v2.16.18.

Based on the release plan published yesterday, the v2.16 is EOL from the 1st of June, we won't like to have instability on that release otherwise we would be forced to keep on maintaining it to make it stable again.

Can you instead repeat the exercise on v3.0? We have at least +6 months on that release.

> Part of that is to upgrade Mina to 2.4 and to make this more clear and in an effort to minimize merge conflicts when/if this is submitted and merge into 3.0, 3.1

Gotcha, so makes more sense in v3.0 which will be maintained a lot longer.

Patch-set: 1
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>