Update patch set 4
Patch Set 4:

> I believe Tony's proposal isn't about limiting the max-width, but rather adding a configurable left/right space based on the max screen size.

Tony changed to max-width and I don't understand why margins would be better.

> By default it would be a noop for Gerrit and for your users.
> Do you see a problem with your user-base? Would they notice any difference?

No. We understand that we could just turn this feature off. But we are generally hesitant to add configuration options for the design/layout of Gerrit, because it can get really difficult to maintain multiple designs at the same time. And we fear such an opt-in feature would typically be only used by one or maybe a few hosts.

And I don't think our users and substantially different than yours. We should be able to design Gerrit for *all* users.
 
> I believe in the example that Tony has shown there is a lot of white space wasted in the current UI. 

Margins are also white space. So the "waste" would be identical.

> The information is spread across the wide screen (left and right), forcing the user to turn the eyes from left to right all the times. Users have raised concerns of tiredness of using it on a daily basis.

I have heard that moving eyes is actually better than staring the entire day.

If users have a problem with that, then they can currently make a choice and make their browser window smaller or zoom in. Both are very decent workarounds. But if we enforce max-width/margin, then the user does not have a choice at all. If they don't like it, they will be stuck and super frustrated.

> Can you propose a way to improve Tony's proposal?

I think the way to go is to actually improve the layout such that information is not pushed completely to the right, etc.

> The health&safety issue raised on the user turning their eyes constantly left and right could be a showstopper.

I don't think this is a health & safety issue and thus not a showstopper.

Patch-set: 4
Reviewer: Gerrit User 1013302 <1013302@173816e5-2b9a-37c3-8a2e-48639d4f1153>