SubmitStrategyListener: Fail for internal errors
If there is an internal error we should throw an exception so that the
user gets a '500 Internal Server Error' response, and not return a
message as part of a '409 Conflict' response that is not meaningful to
the user.

This ensures that these errors are properly counted in our error
metrics.

Throwing an exception also has the advantage that the retry and
auto-tracing logic kicks in.

The RuntimeException's thrown in SubmitStrategyListener are caught in
BatchUpdate#execute where they are wrapped in an UpdateException.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I2962fd872d20e6aad60536985fbc85a5446a7394