Fix commit status when fast-forward on submit is not possible due to conflicts
If FAST_FORWARD_ONLY is configured as submit strategy and submitting a
change series is not possible because fast-forward is not possible, we
set the NOT_FAST_FORWARD commit status only on the tip change of the
change series. Since no commit status is set for the other changes we
return "internal error" for them, which is wrong since this is a normal
situation and not an internal error.

Fix this by setting the NOT_FAST_FORWARD commit status on all changes of
the series if the series cannot be submitted with fast-forward.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I049b3cbe9681ad93fbc4b0e42dd6f7400984045b