Use unique logging tag for submissions
When a change gets submitted we set a logging tag so that all logs that
are written during the submission can be easily found. Before change
Iec8b0db66 the submission ID was a unique string, but now this is no
longer the case. As result of this several submission attempts of the
same change use the same logging tag so that the logs for the different
submission attempts cannot easily be distinguished. Make the logging tag
unique for each request unique again.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I7104086615b4df42d850807f61339a27c71e29e7