Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > Do we need this _and_ the gwtorm update?
> 
> We don't know yet. There are two possibilities:
> 
> * 1. The MySQL driver upgrade fixed the gerrit upgrade problem seen by Eclipse Foundation: then we can preserve bacth operation mode for MySQL dialect in gwtorm and revert my batch mode deactivation change.
> * 2. The MySQL driver upgrade didn't fix gerrit upgrade problem. We need both: disable bacth operation mode for MySQL dialect in gwtorm (may be conditionally?) and the upgrade to the latest patch release of the driver available to avoid further issues.

Is it worth updating the driver to the latest version anyway, regardless of whether it fixes the problem at Eclipse?  In that case we can rebase it on the head of stable-2.16 instead of depending on the gwtorm update.

Patch-set: 3
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>