[DO_NOT_SUBMIT] Add test to reproduce submit error with rebase submit strategies
Tests the following situation:

1. create a change series, consisting out of a merge commit and a normal
   commit
2. before submitting the change series, another non-conflicting change
   gets submitted
3. when the change series gets submitted, Gerrit must perform a
   merge/rebase/cherry-pick

If the submit strategy is 'Rebase Always' or 'Rebase if Necessary' the
submission fails, although the rebase is actually successfully
performed. The submit fails because no commit status is set for the
merge commit.

This change doesn't contain a fix for this issue yet, but only adds a
test that reproduces the issue.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I495ffb7dfcf81c44b4c02f108910020f7450f9ea