Handle merge commits in rebase submit strategies correctly
The rebase submit strategies process commits one by one when a change
series should be integrated (they rebase the changes one by one starting
with the change at the bottom of the change series).

Since we cannot rebase merge changes, it was the intention to merge the
whole series if the change series contained a merge change.

To merge the whole change series we must do:

1. find the minimal set of commits that should be merged (this means we
   exclude commits which will be implicitly integrated by merging the
   series)

2. merge the tip of change series (the tip of the series is the minimal
   set of commits that should be merged, found in step 1.)

Now the code in RebaseSubmitStrategy was always doing 1., but it still
chose the integration Ops as if we would process the commits one by one.
Whether this did the right thing depended on the tip of the change
series. If it was:

* a merge commit, everything worked since we chose the
  RebaseMultipleParentsOp for it which does a merge

* a normal commit, it didn't work since we chose the RebaseOneOp for it
  which cherry-picks the commit on top of the target branch, meaning
  that all other commits in the series are not integrated into the
  target branch (which makes the submit fail because no commit status
  was set for them)

This means:

a) 1. should only be done if we intent to merge the whole change series
   at once

b) if 1. is done we must ensure to do a merge (no matter if the tip of
   the change series is a normal commit or a merge commit)

Now, so far the intention was to merge the whole change series if it
contained any merge commit, but if the merge commits are only at the
bottom of the series doing this is not necessarry, as in this case we
can first merge the merge changes and then rebase the normal changes
that follow. Processing the commits one by one in this case is better
than merging the whole series at once because it allows us to add commit
message footers for the changes that are rebased (which is not possible
if the whole series is merged at once). This is why this change detects
this case and doesn't do a merge of the whole series in this case.

The new test tests the submission of a change series where a normal
change depends on a merge change and hence the whole change series must
be merged at once:

1. it create a change series, consisting out of a merge commit and a
   normal commit
2. before submitting the change series, another non-conflicting change
   gets submitted (to prevent that the fast-forward logic for 'Rebase if
   Necessary' kicks in on submit)
3. when the change series gets submitted, Gerrit must perform a merge of
   the whole series

If the submit strategy was 'Rebase Always' or 'Rebase if Necessary' the
submission in this test failed, because only the normal change was
rebased and the merge change was skipped/ignored. As result of this
there was no commit status for the merge change and hence the submit
request failed.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I495ffb7dfcf81c44b4c02f108910020f7450f9ea