Update patch set 4
Patch Set 4:

What will be the things displayed by default in the change log? Will comments be displayed by default and then when toggling "Show all activity" everything else will be toggled?

Would it lead to cases where users will miss some important information such as Labels and votes?
If comments are displayed by default in the change log, it seems it would be a duplicate of the Comment threads.

Patch-set: 4