Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Should the stable-3.1 branch be merged to master first, before creating stable-3.2 branch?

I tried before and it is way too complex and risky for now: I prefer to get the stable-3.2 cut now, merge the stable-3.1 into stable-3.2, get it well tested and then merge to master.

If we break stable-3.2, we can fix it safely. If we break master, is a lot worse.

Patch-set: 1