Update patch set 1
Patch Set 1:

> > One thing that I struggle to understand is why a DBMS server would close the connection if it is still active. Even if we use a single connection and we do 100s of migration steps for hours, the DB connection should be in constant use, isn't it?
> 
> if the connection is actively used all the time I'd also expect that a wait_timeout wouldn't trigger

But the connection is only used in Schema_146 to scan the accounts as the very first steps, and then never again, possibly for hours. Default wait timeout in SQL server is 8 hours, but every db administrator decreased that default value.

I was able to reproduce the Eclipse Foundation breakage locally: [1] on my test site, without even writing a script to generate 300,000 accounts:

* set up MaridDb
* decrease wait timeout from default 28800 sec to 60 sec:
  
  SET GLOBAL wait_timeout = 60;
  SET GLOBAL interactive_timeout = 60;

* set up 2.14 gerrit site with MySQL database dialect
* add Thread.sleep(120000) to migration data method
* try to migrate to 2.16
* Exactly the same exception stack trace as reported by Eclipse Foundation: [2]. We are hitting wait timeout and the follow-up statement, that is bumping the schema version number, is failing because the connection was already closed

With this discussion, I implemented another version of the fix, that is less intrusive: [3]. It's only firing keep alive database SQL statement pro 1000 accounts (similar to how Git GC is fired) in Schema_146.

As a side effect of this implementation I realized that this change is targeting a wrong branch, because it is only relevant for Schema_146 migration that is of course a part of 2.15. So, if we would like to keep this intrusive approach at all, we should at least move it to stable-2.15 branch. 

[1] http://paste.openstack.org/show/792704
[2] https://bugs.eclipse.org/bugs/show_bug.cgi?id=541221#c60
[3] https://gerrit-review.googlesource.com/c/gerrit/+/264476

Patch-set: 1