Update patch set 6
Patch Set 6:

Does this need to be on stable-2.15?

As far as I know we're not going to make any more 2.15.x releases, so should we just put this on stable-2.16 so it goes into 2.16.19 (the last release before EOL of 2.16.x)?

Patch-set: 6
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>