Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> Does this need to be on stable-2.15?
> 
> As far as I know we're not going to make any more 2.15.x releases, so should we just put this on stable-2.16 so it goes into 2.16.19 (the last release before EOL of 2.16.x)?

This is a critical bug that prevents gerrit users to migrate. However, there is a simple work around: increase wait time configuration option, done.

What we could still do is submit this to stable-2.15, like we did with other changes, but don't conduct a new release from 2.15 branch, but just merge it to stable-2.16 and release 2.16.19.

Patch-set: 6