Optimize latency of diff page load when in background tab
Browser in background tab is running each independant timer no more
than once per second. In diff Gerrit is running rendering in batches
via timers so UI is not stuck. In background tab we don't care if UI
is stuck and timers are penalized so we do work as much synchronous
as possible.

Change-Id: I61dfcca673636c27cecc660bab509de4ee76ccf9