Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I believe we've just cut out a large Gerrit setup in production at Alibaba with this change.
> Running a survey with the community about the ElasticSearch adoption is more than ever a priority before discontinuing more support for older releases.

Totally disagreed and I strongly object to restore support for any discontinued release of ES or for any other third party dependencies for that matter. We have only removed ES releases that reached EOL discontinued similarly, as we removed 2.14 an 2.15 and going to remove 2.16 in 4 weeks, no matter who is using it.

I marked as invalid Lucene releated breakage reported by Eclipse Foundation that is still using gerrit-2.14.20 release. Or are you going to investigate 2.14.20 Lucene breakage (in your spare free time) because Eclipse Foundation is *big* organization?

If anyone is using ES 2, or other releases that reached EOL there are two options:

1. Upgrade to current ES release
2. Buy commercial support

That's how community driven projects are managed.

Patch-set: 2