Add a test for copy detection while diffing patch sets
Adding this test for more visibility of the copy detection in diffs.

Copies are detected when a file is deleted and more than 1 file with the
same content are added. In this case, the added file with the closest
name to the original file is considered a rename and the remaining files
are considered copies.

The implementation of copy detection is part of the RenameDetector
module of JGit.

Change-Id: If69ed04d8d86f21873f0db43e7181e2f2f7d4fee