Update patch set 2
Patch Set 2:

The reason robot comments are not linked correctly is because we skip linking comments (human & robot) to auto-generated messages. This is a new behavior that was introduced in BE. We do it this way for the patch set upload use case: When a user uploads a new patch set, 2 change messages are created (one for PS upload, and one for published comments). Any published comments will be linked to the non auto-generated message - published comments message, hence fixing the empty message problem that appeared in the UI.

I see a small fix in BE for this: robot comments should not ignore auto-generated messages while linking comments to change messages. If this behavior does not contradict with the patch set upload use case then I can implement this small fix today.

Patch-set: 2
Reviewer: Gerrit User 1091982 <1091982@173816e5-2b9a-37c3-8a2e-48639d4f1153>