Add `path` field to every comment and use it when create reply to thread
For renamed files, we support showing comments from before and after,
but with current logic when create a reply, it will always use
the current file path, but when reply comments on left side,
it will use the base patch set that its comparing, and renamed files may
not exist in that patch set.

The fix here is always use first comment's path if exists when adding
a reply, and fallback to thread's path.

Change-Id: If6493eb170c72dae98db8c46cfbebc274d38db78