Change default patchset choice for comments
Clicking on the link to a comment posted on patchset X took user to
Base vs X.
We now take the user to X vs Latest.
A toast informs the user of the new choice and shows the keyboard
shortcut to go to Base vs X.
If user goes to Base vs X (X != latest), then we show a toast telling
users they can go to Base vs Latest via keyboard shortcuts.

Change-Id: Ia2182024917d119b9bf9bd2af8eededf0906b9c5